move schemaSyncer
