+++ dbms/src/Debug/MockTiDB.h
@@ -15,13 +15,13 @@
 #pragma once
 
 #include <Storages/ColumnsDescription.h>
-#include <Storages/Transaction/SchemaSyncer.h>
 #include <Storages/Transaction/TiDB.h>
 #include <Storages/Transaction/Types.h>
+#include <TIDB/Schema/SchemaGetter.h>
 
 #include <atomic>
 
-#include <TIDB/Schema/SchemaGetter.h>
+#include <TIDB/Schema/SchemaSyncer.h>
 
 namespace DB
 {
