+++ dbms/src/Debug/dbgFuncMockTiDBTable.cpp
@@ -24,9 +24,10 @@
 #include <Parsers/ParserRenameQuery.h>
 #include <Parsers/parseQuery.h>
 #include <Storages/Transaction/KVStore.h>
-#include <Storages/Transaction/SchemaSyncer.h>
 #include <Storages/Transaction/TMTContext.h>
 #include <fmt/core.h>
+
+#include <TIDB/Schema/SchemaSyncer.h>
 
 namespace DB
 {
