+++ dbms/src/Flash/BatchCoprocessorHandler.cpp
@@ -18,10 +18,11 @@
 #include <Flash/Coprocessor/InterpreterDAG.h>
 #include <Flash/ServiceUtils.h>
 #include <Storages/IStorage.h>
-#include <Storages/Transaction/SchemaSyncer.h>
 #include <Storages/Transaction/TMTContext.h>
 
 #include <ext/scope_guard.h>
+
+#include <TIDB/Schema/SchemaSyncer.h>
 
 namespace DB
 {
