+++ dbms/src/Interpreters/loadMetadata.cpp
@@ -26,15 +26,15 @@
 #include <Parsers/parseQuery.h>
 #include <Poco/DirectoryIterator.h>
 #include <Poco/FileStream.h>
-#include <Storages/Transaction/SchemaSyncer.h>
 #include <Storages/Transaction/TMTContext.h>
+#include <TIDB/Schema/SchemaNameMapper.h>
 #include <common/ThreadPool.h>
 
 #include <future>
 #include <iomanip>
 #include <thread>
 
-#include <TIDB/Schema/SchemaNameMapper.h>
+#include <TIDB/Schema/SchemaSyncer.h>
 
 
 namespace DB
