+++ dbms/src/Storages/Transaction/ApplySnapshot.cpp
@@ -28,10 +28,11 @@
 #include <Storages/Transaction/Region.h>
 #include <Storages/Transaction/RegionTable.h>
 #include <Storages/Transaction/SSTReader.h>
-#include <Storages/Transaction/SchemaSyncer.h>
 #include <Storages/Transaction/TMTContext.h>
 
 #include <ext/scope_guard.h>
+
+#include <TIDB/Schema/SchemaSyncer.h>
 
 namespace DB
 {
