+++ dbms/src/Storages/Transaction/PartitionStreams.cpp
@@ -26,11 +26,12 @@
 #include <Storages/Transaction/Region.h>
 #include <Storages/Transaction/RegionBlockReader.h>
 #include <Storages/Transaction/RegionTable.h>
-#include <Storages/Transaction/SchemaSyncer.h>
 #include <Storages/Transaction/TMTContext.h>
 #include <Storages/Transaction/TiKVRange.h>
 #include <Storages/Transaction/Utils.h>
 #include <common/logger_useful.h>
+
+#include <TIDB/Schema/SchemaSyncer.h>
 
 namespace DB
 {
