+++ dbms/src/Storages/Transaction/RegionTable.cpp
@@ -20,9 +20,10 @@
 #include <Storages/Transaction/Region.h>
 #include <Storages/Transaction/RegionManager.h>
 #include <Storages/Transaction/RegionTable.h>
-#include <Storages/Transaction/SchemaSyncer.h>
 #include <Storages/Transaction/TMTContext.h>
 #include <Storages/Transaction/TiKVRange.h>
+
+#include <TIDB/Schema/SchemaSyncer.h>
 
 namespace DB
 {
